Split database code
There will also be a `favorite.go` later.
